+++ dpnp/tests/test_functional.py
@@ -1,6 +1,11 @@
 import numpy
 import pytest
-from numpy.testing import assert_array_equal, assert_raises
+from numpy.testing import (
+    assert_array_equal,
+    assert_equal,
+    assert_raises,
+    assert_raises_regex,
+)
 
 import dpnp
 
@@ -65,3 +70,127 @@ class TestApplyOverAxes:
 
         ia = dpnp.arange(24).reshape(2, 3, 4)
         assert_raises(ValueError, dpnp.apply_over_axes, custom_func, ia, 1)
+
+
+class TestPiecewise:
+    def test_simple(self):
+        ia = dpnp.array([0, 0])
+        # Condition is single bool list
+        x = dpnp.piecewise(ia, [True, False], [1])
+        assert_array_equal(x, [1, 0])
+
+        # List of conditions: single bool list
+        x = dpnp.piecewise(ia, [[True, False]], [1])
+        assert_array_equal(x, [1, 0])
+
+        # Conditions is single bool array
+        x = dpnp.piecewise(ia, dpnp.array([True, False]), [1])
+        assert_array_equal(x, [1, 0])
+
+        # Condition is single int array
+        x = dpnp.piecewise(ia, dpnp.array([1, 0]), [1])
+        assert_array_equal(x, [1, 0])
+
+        # List of conditions: int array
+        x = dpnp.piecewise(ia, [dpnp.array([1, 0])], [1])
+        assert_array_equal(x, [1, 0])
+
+        x = dpnp.piecewise(ia, [[False, True]], [lambda x: -1])
+        assert_array_equal(x, [0, -1])
+
+        assert_raises_regex(
+            ValueError,
+            "1 or 2 functions are expected",
+            dpnp.piecewise,
+            ia,
+            [[False, True]],
+            [],
+        )
+        assert_raises_regex(
+            ValueError,
+            "1 or 2 functions are expected",
+            dpnp.piecewise,
+            ia,
+            [[False, True]],
+            [1, 2, 3],
+        )
+
+    def test_two_conditions(self):
+        ia = dpnp.array([1, 2])
+        x = dpnp.piecewise(ia, [[True, False], [False, True]], [3, 4])
+        assert_array_equal(x, [3, 4])
+
+    def test_scalar_domains_three_conditions(self):
+        x = dpnp.piecewise(dpnp.array(3), [True, False, False], [4, 2, 0])
+        assert_equal(x, 4)
+
+    def test_default(self):
+        # No value specified for x[1], should be 0
+        x = dpnp.piecewise(dpnp.array([1, 2]), [True, False], [2])
+        assert_array_equal(x, [2, 0])
+
+        # Should set x[1] to 3
+        x = dpnp.piecewise([1, 2], [True, False], [2, 3])
+        assert_array_equal(x, [2, 3])
+
+    def test_0d(self):
+        x = dpnp.array(3)
+        y = dpnp.piecewise(x, x > 3, [4, 0])
+        assert y.ndim == 0
+        assert y == 0
+
+        x = 5
+        y = dpnp.piecewise(x, [True, False], [1, 0])
+        assert y.ndim == 0
+        assert y == 1
+
+        # With 3 ranges (It was failing, before)
+        y = dpnp.piecewise(x, [False, False, True], [1, 2, 3])
+        assert_array_equal(y, 3)
+
+    def test_0d_comparison(self):
+        x = dpnp.array(3)
+        y = dpnp.piecewise(x, [x <= 3, x > 3], [4, 0])  # Should succeed.
+        assert_equal(y, 4)
+
+        # With 3 ranges (It was failing, before)
+        x = 4
+        y = dpnp.piecewise(x, [x <= 3, (x > 3) * (x <= 5), x > 5], [1, 2, 3])
+        assert_array_equal(y, 2)
+
+        assert_raises_regex(
+            ValueError,
+            "2 or 3 functions are expected",
+            dpnp.piecewise,
+            x,
+            [x <= 3, x > 3],
+            [1],
+        )
+        assert_raises_regex(
+            ValueError,
+            "2 or 3 functions are expected",
+            dpnp.piecewise,
+            x,
+            [x <= 3, x > 3],
+            [1, 1, 1, 1],
+        )
+
+    def test_0d_0d_condition(self):
+        x = dpnp.array(3)
+        c = dpnp.array(x > 3)
+        y = dpnp.piecewise(x, [c], [1, 2])
+        assert_equal(y, 2)
+
+    def test_multidimensional_extrafunc(self):
+        x = dpnp.array([[-2.5, -1.5, -0.5], [0.5, 1.5, 2.5]])
+        y = dpnp.piecewise(x, [x < 0, x >= 2], [-1, 1, 3])
+        assert_array_equal(y, dpnp.array([[-1.0, -1.0, -1.0], [3.0, 3.0, 1.0]]))
+
+    def test_subclasses(self):
+        class subclass(dpnp.ndarray):
+            pass
+
+        x = dpnp.arange(5.0).view(subclass)
+        r = dpnp.piecewise(x, [x < 2.0, x >= 4], [-1.0, 1.0, 0.0])
+        assert_equal(type(r), subclass)
+        assert_equal(r, [-1.0, -1.0, 0.0, 0.0, 1.0])